APIs for Workouts and Movements
Created schemas for workouts and movements. Edited users to contain all
saved workouts as an array. Movements can be retrieved (for searching in
order to create new workouts). The following functionalities can be done
for workouts: get all workouts, get all user-saved workouts, get all
user-created workouts, create a workout, save a pre-existing workout to
the user's profile, and delete a workout from the user's profile. An
authentication middleware was also added to store the user ID from the
JWT.