Scheduling Backend and APIs
Edited the user schema to hold a schedule list. Each day of the week is
an element in the schedule that can hold multiple workouts. However, in
the APIs to add a new workout to a specific day, I restricted each day
to only hold a maximum of 2 workouts. APIs to remove a workout from a
specified day as well as to get the entire schedule were also made.