Logged Workouts -- Backend and APIs
Created a LoggedWorkout schema that holds the original workout, the date
the user completed the workout, total duration of the workout, any notes
the user has to add, and specific details on each movement in the
workout. Each movement has the number of sets and reps/duration that the
user actually did that day (not just what was in the original workout)
and has a new field called highestData where users can input the best
number they got from that day with that movement (e.g. best mile time,
heaviest weight, etc.) APIs to log a workout and get all logged workouts
are included.