Missed Workout Email Reminder
Every day at 8PM, users who have not yet completed their scheduled
workout for that day will be emailed with a reminder. This is completed
using nodemailer and a cron job. The general sendMail.js will allow
other types of emails to be sent in the future if we choose to implement
that feature, such as progress tracking emails.