Logged Workout Analytics Page -- Pre-testing
Initial code for an analytics page for a user's logged workouts. Users
can search for a movement or a workout, and all their logged data
regarding that movement or workout will be displayed in a table sorted
by most recent date. Information on personal records for each type of
movement is also displayed. The user's longest streak of exercising
daily and the user's current streak are displayed as part of the footer.
The page to log a workout has been made while making this page, so I
will next test this analytics page using the logging workouts page.