Logged Workout Analytics -- Post-Testing
Users can search for a movement or a workout on this new page and
receive a table summarizing their logged data for that movement
orworkout. This includes each time they performed that action with the
date, name, sets, reps/duration, and either notes (for workouts) or best
data (for movements). On the right hand side, the current and longest
streaks are displayed above the user's personal records for every
movement they have ever logged on this app. The personal records are
associated with the first date the record was achieved.