Removed mailing feature
The mailing feature only works on fully deployed applications or when
the machine is on during local development. Thus, with our current state
of this application, the automated cron job to send emails to users who
haven't logged their scheduled  workouts at 8PM won't actually run
asynchronously every day at 8PM. This feature is commented out, but the
code is still there for reference and in alignment with our proposal.